add module button permission check

+++ Login/Login/FormPermissionAdd.cs
@@ -51,25 +51,31 @@ namespace Login
         }
         private void buttonAddPermission_Click(object sender, EventArgs e)
         {
+            bool check = User.CheckPermission("addPermission");
             try
             {
-                SqlConnection con2 = new SqlConnection(connection);
-                SqlCommand cmd2 = new SqlCommand("INSERT INTO dbo.g1_permission (permission_name) VALUES (@permissionName);", con2);
-                cmd2.Parameters.AddWithValue("@permissionName", textBox1.Text);
-                con2.Open();
-                int recordsAffected = cmd2.ExecuteNonQuery();
-                if(recordsAffected==1)
+                if (check == true)
                 {
-                    MessageBox.Show("Permission has been added successfully.");
-                    textBox1.Clear();
-                    refreshDataGridView();
+                    SqlConnection con2 = new SqlConnection(connection);
+                    SqlCommand cmd2 = new SqlCommand("INSERT INTO dbo.g1_permission (permission_name) VALUES (@permissionName);", con2);
+                    cmd2.Parameters.AddWithValue("@permissionName", textBox1.Text);
+                    con2.Open();
+                    int recordsAffected = cmd2.ExecuteNonQuery();
+                    if (recordsAffected == 1)
+                    {
+                        MessageBox.Show("Permission has been added successfully.");
+                        textBox1.Clear();
+                        refreshDataGridView();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Error, permission has not been stated.");
+                    }
+
+                    con2.Close();
                 }
                 else
-                {
-                    MessageBox.Show("Error, permission has not been stated.");
-                }
-
-                con2.Close();
+                    MessageBox.Show("You have no permission to add new permission.");
             }
             catch(SqlException ex)
             {
@@ -78,33 +84,39 @@ namespace Login
 
         private void buttonDeletePermission_Click(object sender, EventArgs e)
         {
+            bool check = User.CheckPermission("deletePermission");
             try
             {
-                var selectedItemId = dataGridView1.SelectedCells[0].Value;
-                var selectedItemName = dataGridView1.SelectedCells[1].Value;
-                DialogResult dialogResult = MessageBox.Show("Are you sure you want delete permission: "+selectedItemName+"?", "Delete the permission", MessageBoxButtons.YesNo);
-                if (dialogResult == DialogResult.Yes)
+                if (check == true)
                 {
-                    if (dataGridView1.SelectedRows.Count == 1)
+                    var selectedItemId = dataGridView1.SelectedCells[0].Value;
+                    var selectedItemName = dataGridView1.SelectedCells[1].Value;
+                    DialogResult dialogResult = MessageBox.Show("Are you sure you want delete permission: " + selectedItemName + "?", "Delete the permission", MessageBoxButtons.YesNo);
+                    if (dialogResult == DialogResult.Yes)
                     {
-                        
-                        SqlConnection con3 = new SqlConnection(connection);
-                        SqlCommand cmd3 = new SqlCommand("DELETE FROM dbo.g1_permission WHERE id_permission=@idPermission;", con3);
-                        cmd3.Parameters.AddWithValue("@idPermission", selectedItemId);
-                        con3.Open();
-                        cmd3.ExecuteNonQuery();
-                        con3.Close();
-                        MessageBox.Show("The permission with id: " + selectedItemId + ", has been deleted.");
-                        refreshDataGridView();
+                        if (dataGridView1.SelectedRows.Count == 1)
+                        {
+
+                            SqlConnection con3 = new SqlConnection(connection);
+                            SqlCommand cmd3 = new SqlCommand("DELETE FROM dbo.g1_permission WHERE id_permission=@idPermission;", con3);
+                            cmd3.Parameters.AddWithValue("@idPermission", selectedItemId);
+                            con3.Open();
+                            cmd3.ExecuteNonQuery();
+                            con3.Close();
+                            MessageBox.Show("The permission with id: " + selectedItemId + ", has been deleted.");
+                            refreshDataGridView();
+                        }
+                        else
+                            MessageBox.Show("Error, the permission has not been deleted.");
                     }
-                    else
-                        MessageBox.Show("Error, the permission has not been deleted.");
+                    else if (dialogResult == DialogResult.No)
+                        return;
                 }
-                else if (dialogResult == DialogResult.No)
-                    return;
+
+                else
+                    MessageBox.Show("You have no permission to delete a permission");
+
             }
-                
-                
 
             catch (SqlException ex)
             {
